+++ src/chartreuse/tests/unit_tests/test_alembic.py
@@ -17,7 +17,7 @@ def test_respect_not_empty_database(mocker):
     """
     table_list = ["alembic_version", "foobar"]
     mocker.patch("chartreuse.utils.AlembicMigrationHelper._get_table_list", return_value=table_list)
-    mocker.patch("chartreuse.utils.AlembicMigrationHelper._get_alembic_current", return_value='123')
+    mocker.patch("chartreuse.utils.AlembicMigrationHelper._get_alembic_current", return_value="123")
     alembic_migration_helper = chartreuse.utils.AlembicMigrationHelper("foo", configure=False)
     assert alembic_migration_helper.is_postgres_empty() is False
 
@@ -113,9 +113,11 @@ def test_additional_parameters(mocker):
 
     mocked_run_command = mocker.patch("chartreuse.utils.alembic_migration_helper.run_command")
 
-    alembic_migration_helper = chartreuse.utils.AlembicMigrationHelper("foo", configure=False, additional_parameters="foo bar")
+    alembic_migration_helper = chartreuse.utils.AlembicMigrationHelper(
+        "foo", configure=False, additional_parameters="foo bar"
+    )
     alembic_migration_helper.upgrade_db()
-    mocked_run_command.assert_called_with("alembic foo bar upgrade head", cwd='/app/alembic')
+    mocked_run_command.assert_called_with("alembic foo bar upgrade head", cwd="/app/alembic")
 
 
 def test_additional_parameters_current(mocker):
@@ -126,6 +128,8 @@ def test_additional_parameters_current(mocker):
     mocked_run_command = mocker.patch("chartreuse.utils.alembic_migration_helper.run_command")
     mocked_run_command.return_value = ("bar", None, 0)
 
-    alembic_migration_helper = chartreuse.utils.AlembicMigrationHelper("foo", configure=False, additional_parameters="foo bar")
+    alembic_migration_helper = chartreuse.utils.AlembicMigrationHelper(
+        "foo", configure=False, additional_parameters="foo bar"
+    )
     alembic_migration_helper._get_alembic_current()
-    mocked_run_command.assert_called_with("alembic foo bar current", cwd='/app/alembic', return_result=True)
+    mocked_run_command.assert_called_with("alembic foo bar current", cwd="/app/alembic", return_result=True)