+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # ChangeLog
 
-# 3.0.1 (2020-04-13)
+# 3.1.0 (2021-05-03)
+## Feature
+- Add support for PG clusters managed by postgres-operator (Patroni PG):
+  - Wait for the PG cluster to be configured before running the migrations. (See [Note 1.](https://gitlab.wiremind.io/wiremind/devops/chartreuse/-/tree/v3.1.0#notes) for more details)
+  - When Alembic run migrations against a Patroni PG, the owner Role should be set. (See [Note 1.](https://gitlab.wiremind.io/wiremind/devops/chartreuse/-/tree/v3.1.0#notes) for more details)
+
+# 3.0.1 (2021-04-13)
 ## Fix
 - ensure_safe_run: Support more sem versions.
 - Upgrade dependencies
 
-# 3.0.0 (2020-02-17)
+# 3.0.0 (2021-02-17)
 ## BREAKING CHANGE
 - Check the compatibility between Chartreuse's python package version and Chartreuse's Helm Chart version. Raise if the versions are not compatible.
 - helm-chart: Deprecate `.Values.runMigrationInPreDeployment`, use `.Values.upgradeBeforeDeployment` instead.